requirements.txt
beautifulsoup4
Scrapy
selenium
webdriver-manager
pandas
sentence-transformers
tqdm
langchain-community
langchain
langchain-huggingface
langchain_huggingface
jq
streamlit 
faiss-cpu 
openai
accelerate 
bitsandbytes
dotenv